aws/requirements.txt
octotools
openai
flotorch-core
streamlit
python-dotenv